views/network: fix focus handling
Now when the view appears on the display, the focus is set on the 'Done'
button (because focus_buttons == True and _hostname_form validates).

And when we add/remove an interface, the focus goes back to the iface
selector. This helps if, for some, reason the actions widget below
iface_selector contains no widgets capable of receiveing a focus.