StorageInstaller: deactivate block devices on cleanup
It's a good citizen's practice to clean up after themselves,
so we closing/deactivating all 'complex' block devices (LVMs,
RAIDs, crypto volumes) after we are done with the installation,
so their presence will not block any further work with underlying
block devices on the host.

For example, it helps if one decides to repeat the installation
using a (possibly adjusted) yaml config file.