Execute run_cmd* commands in a separate process group
Both the installer process and a run_cmd-started process
are put into one process group (foreground group). This means
when we send a tty signal (such as SIGINT) it's delivered
to both the processes. As a result, despite the fact that
SIGINT and SIGQUIT are ignored in the installer process
when operating in the iso mode, they still get delivered
to the run_cmd process, which, in turn, terminates.

The proposal is to put the run_cmd process in a separate
process group to avoid delivery of tty signals to it.

The only noticeable drawback of this approach is that
when we are operating _not_ in the iso mode and press Ctrl+C
while a long-running run_cmd process is being executed,
this process is left running until it is either completed
or terminated by other methods.