installers/kernel: modify /etc/default/grub
* Add GRUB_DEFAULT=saved so that we can use grub-set-default.

* Enable os-prober by default:

This is to disable the annoying warning every time the new kernel
is installed and/or initramfs updated:

    Warning: os-prober will not be executed to detect other bootable partitions.
    Systems on them will not be added to the GRUB boot configuration.
    Check GRUB_DISABLE_OS_PROBER documentation entry.

For example Fedora reverted the commits disabling os-prober by default.
We can do the same, but with the option in /etc/default/grub, so it is
visible for the user and it can be easily changed if needed.

os-prober was disabled by default for security reasons, but 30_os-prober
hook is no-op when os-prober package has not been installed and this is
the default in Alpaca.

Moreover, when os-prober is installed manually, grub will do print the
warning, which is more noticable if there were no warnings before:

    Warning: os-prober will be executed to detect other bootable partitions.