NetworkInstaller: interfaces on the target may differ
NetworkInstaller had an assumption that interface names on
the target are the same as on the host. And this is not the
case when we run the installer under Ubuntu to perform
an _automated_ installation of Alpaquita, as interface names
in Ubuntu and Alpaquita are different.

To address that we now construct the NetworkManager's state by
reading the data right from the config without looking at the
host network interfaces.