StorageManager: use parted to inform the OS of new partitions
Valery Ushakov (valery.ushakov@bell-sw.com) proposed to consider
using parted as it better informs the OS of created/deleted
partitions. In particular, parted unlike 'blockdev --rereadpt'
works fine with loopback block devices.

So now we fist create partitions with sfdisk (as it doesn't require
manual calculation of offsets and alignments) and then 'properly'
recreate them with parted. In my testing this approach
always resulted in correct creation/deletion of parition block devices.

In the future, it would be better to convert the existing sfdisk
code to parted.