Pass --keys-dir with each apk command
The existing approach of passing --keys-dir only when
we install distro-keys does not work for installations
with repos signed with custom keys.

One possible fix for that is to always pass --keys-dir.
And the user will be responsible that the --keys-dir directory
contains all necessary keys. In the case when we combine
distro and custom repos the --keys-dir must contain both
distro and custom keys. This is actually in line with our
documentation (AUTOMATING_INSTALLATION.md).

Implementing this fix required moving apk calls from installers
to a separate entity. It also allowed moving other apk logic
there (such as creation of the repositories file).