BaseInstallerController: create/remove TARGET_ROOT
For odd reasons the test environment contained /mnt/target_root
precreated. This hid the bug that

  apk.root_dir = self.TARGET_ROOT

is executed before self.TARGET_ROOT is created. Sic.

To fix that I'm proposing to handle the creation and deletion
of self.TARGET_ROOT right in the BaseInstallerController. Having
a separate Installer for that would be an overkill.