Install distro-base before other packages
It was found that when we install bash along with distro-base,
it may happen that bash gets installed before distro-base, and
/etc/shells becomes inconsistent. This situation may be fixed
by making bash depend on distro-base. However, as distro-base
is really "base" and is present in every instance, a different
solution was proposed - just install it before any other packages.

Plus now we always pass --keys-dir to apk, there is no need
to install distro-keys separately from other packages, as it's
already a dependency for distro-base.