Always create /boot in UI installations
If the following conditions are met:
 * there is no seperate /boot
 * / is on xfs
 * we've just upgraded musl and kernel
 * we've just issued a reboot
then busybox init may fail to correctly umount -o ro / during
the reboot procedure. This may make updated files in /boot corrupted,
and grub may not be able to load them on the next boot.

To avoid this situation we always create a separate /boot.

Plus reduced the size of /boot from 1G to 512MB. It should be enough.